+++ CHANGES.rst
@@ -31,7 +31,7 @@ Unreleased
  * **[CHANGE]** Fixed the `subreddits` attribute of
    :class:`praw.objects.Multireddit` being returned as a list of dicts.
    It is now a list of Subreddit objects.
- * **[FEATURE]** Addded :meth:`get_comment_replies` and
+ * **[FEATURE]** Added :meth:`get_comment_replies` and
    :meth:`get_post_replies` to the :class:`praw.__init__.PrivateMessagesMixin`
 
 PRAW 3.0.0